requirements.txt
aiosmtpd==1.4.4.post2
atpublic==3.1.1
attrs==22.2.0
certifi==2022.12.7
click==8.0.4
colorama==0.4.6
Flask==2.2.2
greenlet==2.0.1
importlib-metadata==6.0.0
itsdangerous==2.0.1
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-connector-python==8.0.32
pip==23.0.1
protobuf==3.20.3
setuptools==65.6.3
SQLAlchemy==1.4.39
Werkzeug==2.2.2
wheel==0.38.4
wincertstore==0.2
zipp==3.11.0